Fixed integration environment
Transbank requires a second step to reach production: use the integration environment and endpoints but with custom certificates. This adds the possibility to use these certificates when not in production (yet). For this, a valid Commerce Code is needed, otherwise there is no point in having valid certificates but no Commerce Code.